Update python 2.4.py

+++ python 2.4.py
@@ -1,9 +1,9 @@
 #Find greatest of 3 number entered by user
 
 
-num1=int(input("Enter first number"))
-num2=int(input("Enter second number"))
-num3=int(input("Enter third number"))
+num1=float(input("Enter first number"))
+num2=float(input("Enter second number"))
+num3=float(input("Enter third number"))
 if(num1>num2 and num1>num3):
     print(num1,"is the biggest among the three numbers")
 elif(num2>num1 and num2>num3):
